backend/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==23.2.1
httpx>=0.26.0,<0.27.2

# OpenAI API
openai==1.12.0

# Supabase
supabase==2.14.0

# Audio processing
ffmpeg==1.4  # Python wrapper for ffmpeg